Backend/requirements.txt
langchain
langchain-groq
python-dotenv
PyPDF2